Implement maze generation algorithm and elaborate cell properties
Upgraded the implementation of the Maze class, which now can generate mazes using a randomised algorithm. This involved elaborating properties of the Cell class. The new approach ensures every Cell object knows if it has been visited during the maze generation and refined the logic for drawing cell walls; this enhances the performance of maze creation. A seeding option for the random number generator was added, enabling the creation of identical mazes for debugging.